Gate to secure routes

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -14,6 +14,9 @@ class AttendeeController extends Controller
     use CanLoadRelationships;
 
     private array $relations = ['user'];
+    public function __construct(){
+        $this->middleware('auth:sanctum')->except(['index', 'show' , 'update']);
+    }
 
     public function index(Event $event)
     {
@@ -50,8 +53,15 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+
+        if (\Illuminate\Support\Facades\Gate::denies('delete-attendee',  [$event , $attendee])) {
+            return response()->json([
+                'message' => 'You are not autrhorized to delete this attendee !'
+            ], 403);
+        }
+
         $attendee->delete();
 
         return response(status: 204);
